Modified helper for constructTree to take depth.  Also added method in Player to get a list of all players at a position for a season in order of Fantasy Points scored.

+++ src/groovy/com/traderapist/draft/MinimaxTreeNode.groovy
@@ -13,7 +13,7 @@ class MinimaxTreeNode {
 	/**
 	 * The perceived value of this state to the owner we're creating the tree for.
 	 */
-	Double value
+	Short value
 
 	/**
 	 * The parent node in the tree.
